seperated different things into different files and made year automatically detected for the aoc command
