+++ modules/aoc/aocCommand.ts
@@ -1,7 +1,8 @@
 import {ISlashCommand} from "../../types";
 import {CommandInteraction} from "discord.js";
-import {embedLeaderboard} from "./aocLeaderboardParser";
+import {embedLeaderboard} from "./aocEmbedGenerator";
 import {get} from "../../lib/configmanager";
+import {DateTime} from "luxon";
 
 /**
  * Slash command definition for /aoc, a command that sends the current leaderboard
@@ -12,7 +13,8 @@ export default {
         description: "Returns the Advent of Code Leaderboard"
     },
     handler: async function(interaction: CommandInteraction) {
-        const embed = await embedLeaderboard(get("id", "aoc") as number, get("year", "aoc") as number);
+        const now = DateTime.now().setZone("Europe/Berlin");
+        const embed = await embedLeaderboard(get("id", "aoc") as number, now.year);
         await interaction.reply({ embeds: [embed] });
     }
 } as ISlashCommand;
